Manually test for and build asr.so when it's missing
Using the normal dependency rules causes make to try to
rebuild asr.so when asr.zip is missing, causing a build failure on
the pi. I could not find a way to turn of this behavior. Manually
testing for the existence of asr.so and only rebuilding when its
not there makes it possible to run the packaging job automatically.